Create README.md
The first reason I created this project is to be more familiar with Git commands and to discover the Git-Hub plateform. I don't really know how this project will grow because I have others projects on my road map ! Undoubtly, increase my skill in Java with the EE edition is my main aim to join a professional team in development  of software and services in the west of France were I live.